--- src/AssetPreviewProps.ts
@@ -1,7 +1,0 @@
-import {ReactNode} from 'react'
-import {GoogleDriveFileAssetValue} from './GoogleDriveFileAsset'
-
-export interface AssetPreviewProps {
-  value?: GoogleDriveFileAssetValue | null
-  children?: ReactNode
-}
