+++ src/components/GoogleDriveFileInput.tsx
@@ -2,9 +2,8 @@
 import React, {useCallback, useMemo, useRef} from 'react'
 import useDrivePicker from 'react-google-drive-picker'
 import {Stack, Button, Grid, Card, Text} from '@sanity/ui'
-import type {GoogleDriveFileInputProps} from './googleDriveFileInputProps'
+import type {GoogleDriveFileAssetValue, GoogleDriveFileInputProps} from '../types'
 import {PatchEvent, set, unset} from 'sanity'
-import type {GoogleDriveFileAssetValue} from './GoogleDriveFileAsset'
 import {AssetEmbed} from './AssetEmbed'
 import {AssetPreview} from './AssetPreview'
 import type {PickerCallback} from 'react-google-drive-picker/dist/typeDefs'
