--- src/googleDrivePluginConfig.ts
@@ -1,6 +1,0 @@
-// eslint-disable-next-line no-warning-comments
-// TODO: add OAuth token as configuration
-export interface GoogleDrivePluginConfig {
-  clientId: string
-  apiKey: string
-}
