+++ src/helpers/isGoogleDriveFileInputProps.ts
@@ -1,6 +1,6 @@
 import {type InputProps, isObjectInputProps} from 'sanity'
-import {FILE_ASSET} from './fileAsset'
-import type {GoogleDriveFileInputProps} from './googleDriveFileInputProps'
+import type {GoogleDriveFileInputProps} from '../types'
+import {FILE_ASSET} from '../schema'
 
 export function isGoogleDriveFileInputProps(props: InputProps): props is GoogleDriveFileInputProps {
   return isObjectInputProps(props) && props.schemaType.type?.name === FILE_ASSET
